requirements.txt
#Python 依赖列表
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
openai>=1.40.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
